+++ packages/amplify-cli/src/index.ts
@@ -17,7 +17,6 @@ import { isCI } from 'ci-info';
 import { EventEmitter } from 'events';
 import * as fs from 'fs-extra';
 import * as path from 'path';
-
 import { formatter, printer, prompter } from 'amplify-prompts';
 import { logInput } from './conditional-local-logging-init';
 import { attachUsageData, constructContext } from './context-manager';