+++ packages/amplify-e2e-core/src/export/index.ts
@@ -3,7 +3,7 @@ import { nspawn as spawn, getCLIPath } from '..';
 export function exportBackend(cwd: string, settings: { exportPath: string }): Promise<void> {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['export', '--out', settings.exportPath], { cwd, stripColors: true })
-      .wait('For more information: docs.amplify.aws/cli/usage/export-to-cdk')
+      .wait('For more information: https://docs.amplify.aws/cli/usage/export-to-cdk')
       .sendEof()
       .run((err: Error) => {
         if (!err) {