+++ packages/amplify-cli-npm/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [9.0.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/cli@8.5.1...@aws-amplify/cli@9.0.0) (2022-06-22)
+
+
+
+# 9.0.0 (2022-06-21)
+
+
+* feat!: breaking change (#10628) ([7fadbf5](https://github.com/aws-amplify/amplify-cli/commit/7fadbf53b8dc89d768779fb588971f5c5b6e70d7)), closes [#10628](https://github.com/aws-amplify/amplify-cli/issues/10628)
+
+
+### BREAKING CHANGES
+
+* this is a breaking change
+
+
+
+
+
+## [8.5.1](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/cli@8.5.0...@aws-amplify/cli@8.5.1) (2022-06-15)
+
+**Note:** Version bump only for package @aws-amplify/cli
+
+
+
+
+
+# [8.5.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/cli@8.4.0...@aws-amplify/cli@8.5.0) (2022-06-13)
+
+
+### Features
+
+* no op commit ([#10553](https://github.com/aws-amplify/amplify-cli/issues/10553)) ([4e16f6b](https://github.com/aws-amplify/amplify-cli/commit/4e16f6b5d0c488a9357c3dbca11095bd6c4419b5))
+
+
+
+
+
 # [8.4.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/cli@8.3.1...@aws-amplify/cli@8.4.0) (2022-06-03)