+++ packages/amplify-cli/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [9.0.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/cli-internal@8.5.1...@aws-amplify/cli-internal@9.0.0) (2022-06-22)
+
+
+
+# 9.0.0 (2022-06-21)
+
+
+* feat!: breaking change (#10628) ([7fadbf5](https://github.com/aws-amplify/amplify-cli/commit/7fadbf53b8dc89d768779fb588971f5c5b6e70d7)), closes [#10628](https://github.com/aws-amplify/amplify-cli/issues/10628)
+
+
+### BREAKING CHANGES
+
+* this is a breaking change
+
+
+
+
+
+## [8.5.1](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/cli-internal@8.5.0...@aws-amplify/cli-internal@8.5.1) (2022-06-15)
+
+**Note:** Version bump only for package @aws-amplify/cli-internal
+
+
+
+
+
+# [8.5.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/cli-internal@8.4.0...@aws-amplify/cli-internal@8.5.0) (2022-06-13)
+
+
+
+# 8.5.0 (2022-06-10)
+
+
+### Bug Fixes
+
+* create same identifier for flow data and reporter ([#10516](https://github.com/aws-amplify/amplify-cli/issues/10516)) ([ccd212e](https://github.com/aws-amplify/amplify-cli/commit/ccd212eaef9890c03435ad5244f16ccc4db4fe22))
+* headless diagnose prompt ([#10557](https://github.com/aws-amplify/amplify-cli/issues/10557)) ([b4ca7fd](https://github.com/aws-amplify/amplify-cli/commit/b4ca7fd24debc1a67aace6eabcd2ee35f20bde11))
+* revert running amplify version on install ([#10529](https://github.com/aws-amplify/amplify-cli/issues/10529)) ([1539b24](https://github.com/aws-amplify/amplify-cli/commit/1539b246b04ae7f68dc83522c3ff4df96e602c08))
+
+
+### Features
+
+* implemented push normalization ([#10469](https://github.com/aws-amplify/amplify-cli/issues/10469)) ([e70d6c9](https://github.com/aws-amplify/amplify-cli/commit/e70d6c95cd9826a9d2785607341e8e1432306a30))
+
+
+
+
+
 # [8.4.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/cli-internal@8.3.1...@aws-amplify/cli-internal@8.4.0) (2022-06-03)