+++ packages/amplify-util-uibuilder/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.30](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-util-uibuilder@1.2.29...@aws-amplify/amplify-util-uibuilder@1.2.30) (2022-06-22)
+
+**Note:** Version bump only for package @aws-amplify/amplify-util-uibuilder
+
+
+
+
+
+## [1.2.29](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-util-uibuilder@1.2.28...@aws-amplify/amplify-util-uibuilder@1.2.29) (2022-06-15)
+
+**Note:** Version bump only for package @aws-amplify/amplify-util-uibuilder
+
+
+
+
+
+## [1.2.28](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-util-uibuilder@1.2.27...@aws-amplify/amplify-util-uibuilder@1.2.28) (2022-06-13)
+
+**Note:** Version bump only for package @aws-amplify/amplify-util-uibuilder
+
+
+
+
+
 ## [1.2.27](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-util-uibuilder@1.2.26...@aws-amplify/amplify-util-uibuilder@1.2.27) (2022-06-03)
 
 **Note:** Version bump only for package @aws-amplify/amplify-util-uibuilder